requirements.txt
ebooklib>=0.18
beautifulsoup4>=4.12
lxml>=5.0
snowflake-connector-python>=3.6
cryptography>=42.0
requests>=2.31
